old_talk_model/pyproject.toml
[tool.poetry]
name = "pymdp_talk"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11,<3.12"
inferactively-pymdp = "^0.0.7.1"
numpy = "^1.26.0"
matplotlib = "^3.8.0"
seaborn = "^0.12.2"
python-dotenv = "^1.0.0"
openai = "^0.28.1"
h5py = "^3.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
